+++ src/main/java/com/pe/Bruma/pedido/entity/Pedido.java
@@ -8,7 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -19,7 +18,7 @@ import java.util.List;
 public class Pedido {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger pedido_id;
+    private Long pedido_id;
     private LocalDateTime fecha_pedido=LocalDateTime.now();
     private BigDecimal total;
     private String estado;
